% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_structures.R
\name{include_structure}
\alias{include_structure}
\title{Include a protein structure}
\usage{
include_structure(pdb_id)
}
\arguments{
\item{pdb_id}{Protein Database ID of the protein structure to embed}
}
\value{
Code that is inserted into the output document
}
\description{
Use this function to embed a protein structure into your html output document.
}
