% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_structures.R
\name{init_molstar}
\alias{init_molstar}
\title{Initialize protein structure display}
\usage{
init_molstar()
}
\value{
Code that is inserted into the output document
}
\description{
Put this at the top of your Rmarkdown document.
}
